scripts/requirements.txt
kafka-python>=2.0.0
aws-msk-iam-sasl-signer-python>=1.0.2
boto3>=1.26.0
requests>=2.28.0